eventlistener is registered/unregistered only once
